WIP on generic lifecycle fn implementation
